% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{correctTimes}
\alias{correctTimes}
\title{timezone}
\usage{
correctTimes(df, time_col = "DateTime_UTC", tz_offset)
}
\arguments{
\item{df}{A data frame object}

\item{time_col}{Column with timestamps. Defaults to "DateTime_UTC"}

\item{tz_offset}{Number of hours different from UTC to local time. Note whether positive or negative.}
}
\value{
Adds three timestamp columns to the original data frame argument
}
\description{
Converts UTC timestamps to DateTime, Date, and Time columns in local time
}
\examples{
matched_detects <- timezone(matched_detects, tz_offset = 6)
}
