% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{deduplicate}
\alias{deduplicate}
\title{deduplicate}
\usage{
deduplicate(
  df,
  fish_col = "FishID",
  time_col = "DateTime_LT",
  threshold = 60,
  type = "hammer"
)
}
\arguments{
\item{df}{A data frame containing receiver detections}

\item{fish_col}{The name (in quotes) of the column that contains individual animal IDs}

\item{time_col}{The name (in quotes) of the column that contains date-times}

\item{threshold}{Value for how far apart detections must be to be considered *not* duplicates}

\item{type}{One of either "hammer," which will just hammer out all detections within the interval, or "spear," which will do so in a step-wise way to conserve as many detections as possible given the threshold.}
}
\value{
A data frame with only detections that meet the threshold for inclusion
}
\description{
Removes duplicate detections from receiver data.
}
