% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthlengths.R
\name{depthlengths}
\alias{depthlengths}
\title{depthlengths}
\usage{
depthlengths(
  df,
  fish_col,
  time_col = "DateTime_LT",
  depth_col,
  skip.na = FALSE
)
}
\arguments{
\item{df}{A data frame object}

\item{fish_col}{Column with individual tag identifiers}

\item{time_col}{Column with timestamps. Defaults to "DateTime_LT"}

\item{depth_col}{Column with depth values}

\item{skip.na}{Boolean value for whether to skip NA values. Defaults to FALSE: if the number of NA depths is > 1, function will stop.}
}
\value{
Adds diffsecs and diffdepth columns to original dataframe
}
\description{
Adds columns for difference in seconds and difference in depth to a time series of depth readings. Note that this function will assume anything passed as depth_col contains only depths; if, for example, it receives a SensorValue column, it is the coder's responsibility to interpret mixed sensor values accordingly.
}
\examples{
vertical_rates <- depthlengths(vr2_data, fish_col = "FishID", time_col = "DateTime_LT", depth_col = "DepthData")
}
