% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{diagnose}
\usage{
diagnose(
  df,
  time_col = DateTime_LT,
  fish_col,
  sensors = FALSE,
  sensor_col = NA,
  sensor_vals = NA
)
}
\arguments{
\item{df}{A data frame object}

\item{time_col}{Name (*not* in quotes) of the time column in the data frame. Must be POSIX. Defaults to DateTime_LT.}

\item{fish_col}{Name (*not* in quotes) of the column with fish names in data frame.}

\item{sensors}{Boolean value related to whether the transmitters have sensors. Defaults to FALSE.}

\item{sensor_col}{Name (*not* in quotes) of the column that indicates sensor type. Defaults to NA.}

\item{sensor_vals}{Name (*not* in quotes) of the column with sensor values. Defaults to NA.}
}
\value{
Suite of plots looking at the number of VR2 detections over the time period.
}
\description{
Produces basic plots looking at the number of detections for VR2 data.
}
