% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateDates.R
\name{interpolateDates}
\alias{interpolateDates}
\title{interpolateDates}
\usage{
interpolateDates(
  df,
  time_col,
  depth_col,
  var = "Temp",
  interval = "day",
  supplied.date = FALSE
)
}
\arguments{
\item{df}{A data frame object. Currently, must contain a date column and a depth column.}

\item{time_col}{Name (in quotes) of the time column in the data frame. Should be POSIX (see supplied.date if not).}

\item{depth_col}{Name (in quotes) of the column with depths}

\item{var}{Variable of interest. Defaults to "Temp"}

\item{interval}{Temporal resolution of interpolated data frame. Defaults to "day"}

\item{supplied.date}{Whether to convert time_col to POSIX. Set to TRUE when time_col is supplied as a Date ("%Y-%m-%d"). Defaults to FALSE.}
}
\value{
Interpolated data for the variable of interest.
}
\description{
Interpolates temperature data to populate missing dates at the depth values for which data exist.
}
\examples{
interpolated_temps <- interpolateDates(temp_profiles, time_col = "Date_LT", depth_col = "Depths", interval = "day")
}
