% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateDepths.R
\name{interpolateDepths}
\alias{interpolateDepths}
\title{interpolateDepths}
\usage{
interpolateDepths(
  df,
  var = "Temp",
  time_col,
  depth_col,
  min,
  max,
  interval,
  pos.con = "\%Y-\%m-\%d",
  sinkends = FALSE
)
}
\arguments{
\item{df}{A data frame object}

\item{var}{Variable of interest. Defaults to "Temp"}

\item{time_col}{Name (in quotes) of the time column in the data frame. Must be POSIX.}

\item{depth_col}{Name (in quotes) of the column with depths}

\item{min}{Minimum depth up to which to interpolate}

\item{max}{Maximum depth down to which to interpolate}

\item{interval}{Resolution (in metres) of interpolated data frame.}

\item{pos.con}{Configuration of POSIX object. Defaults to "%Y-%m-%d"}

\item{sinkends}{Boolean value related to whether data should be "sunk" to min and max values even if there are no measured values on which to base the interpolation.}
}
\value{
Interpolated data for the variable of interest.
}
\description{
Interpolates temperature data to populate missing depths at the date values for which data exist.
}
\examples{
interpolated_temps <- interpolateDepths(temp_profiles, time_col = "Date_LT", depth_col = "Depths", var = "Temp",
 interval = 0.1, min = 0, max = 15, sinkends = TRUE)
}
