% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepFish.R
\name{keepFish}
\alias{keepFish}
\title{keepFish}
\usage{
keepFish(df, fish_col, fishnames)
}
\arguments{
\item{df}{A data frame object}

\item{fish_col}{The name (in quotes) of the column that contains individual animal IDs}

\item{fishnames}{Vector of known IDs that should be kept}
}
\value{
A data frame that only contains detections that match a known ID
}
\description{
Filters detection data frame to include only those that match known IDs
}
\examples{
matched_detects <- keepFish(vr2_detects, fish_col = "FishID", fishnames = fish_surgery$FishID)
}
