% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDetectsFolder.R
\name{readDetectsFolder}
\alias{readDetectsFolder}
\title{readDetectsFolder}
\usage{
readDetectsFolder(foldername)
}
\arguments{
\item{foldername}{Path to the folder containing .csv files}
}
\value{
A data frame comprised of all the .csv files bound together
}
\description{
Reads in .csv files from a single folder and compiles them into one long data frame
}
\examples{
vr2_detects <- readDetectsFolder("C:/Users/janefish/Field Data/VR2 Download 20250814/csv files")
}
