% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\title{season}
\usage{
season(df, date_col = "Date_LT", type = "ice", firstdays, hemisphere = "north")
}
\arguments{
\item{df}{A data frame object with detections}

\item{date_col}{The name (in quotes) of the column that contains dates. Defaults to "Date_LT"}

\item{type}{One of "ice" or "month." If "ice," supply a vector of dates; if "month," supply a vector of numeric values corresponding to months.}

\item{firstdays}{Vector of first days corresponding to each season.}

\item{hemisphere}{Defaults to "north" - specify as "south" to flip the order of the seasons.}
}
\value{
Adds a "Season" column to the original data frame supplied.
}
\description{
Assign seasons to detection data based on first-day thresholds
}
\examples{
matched_detects <- season(matched_detects,
  date_col = "Date_LT", type = "ice",
  firstdays = c("2025-03-14", "2025-05-25", "2025-10-02", "2025-11-14"), hemisphere = "north"
)
}
