% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sundial.R
\name{sundial}
\alias{sundial}
\title{sundial}
\usage{
sundial(
  df,
  date_col,
  time_col,
  latitude,
  longitude,
  mindate = min(df[[date_col]]),
  maxdate = max(df[[date_col]]),
  timezone
)
}
\arguments{
\item{df}{A data frame object.}

\item{date_col}{Name (in quotes) of the column with dates in the data frame.}

\item{time_col}{Name (in quotes) of the column with times.}

\item{latitude}{Latitude to use for calculating sunrise/sunset.}

\item{longitude}{Longitude to use for calculating sunrise/sunset.}

\item{mindate}{First date for which to calculate. Defaults to minimum value of date_col}

\item{maxdate}{Last date for which to calculate. Defaults to maximum value of date_col}

\item{timezone}{Timezone associated with df}
}
\value{
Adds columns to input df reflecting sunrise, sunset, and period of the day.
}
\description{
Applies suncalc::getSunlightTimes and then derives periods of the day based on sunrise/sunsets.
}
\examples{
vr2_detects_daynight <- sundial(vr2_detects, date_col = "Date_LT", time_col = "Time_LT", latitude = -36, longitude = 174,
timezone = "NZ")
}
