% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateDates.R
\name{interpolateDates}
\alias{interpolateDates}
\title{interpolateDates}
\usage{
interpolateDates(df, date_col, depth_col, var = "Temp", interval = "day")
}
\arguments{
\item{df}{A data frame object. Currently, must contain a date column and a depth column.}

\item{date_col}{Name (in quotes) of the column with dates in the data frame.}

\item{depth_col}{Name (in quotes) of the column with depths}

\item{var}{Variable of interest. Defaults to "Temp"}

\item{interval}{Temporal resolution of interpolated data frame. Defaults to "day"}
}
\value{
Interpolated data for the variable of interest.
}
\description{
Interpolates temperature data to populate missing dates at the depth values for which data exist.
}
\examples{
interpolated_temps <- interpolateDates(temp_profiles, date_col = "Date_LT", depth_col = "Depths", interval = "day")
}
