% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateDepths.R
\name{interpolateDepths}
\alias{interpolateDepths}
\title{interpolateDepths}
\usage{
interpolateDepths(
  df,
  var = "Temp",
  date_col,
  depth_col,
  min,
  max,
  interval,
  sinkends = FALSE
)
}
\arguments{
\item{df}{A data frame object}

\item{var}{Variable of interest. Defaults to "Temp"}

\item{date_col}{Name (in quotes) of the column with dates in the data frame.}

\item{depth_col}{Name (in quotes) of the column with depths}

\item{min}{Minimum depth up to which to interpolate}

\item{max}{Maximum depth down to which to interpolate}

\item{interval}{Temporal resolution of interpolated data frame. Defaults to "day"}

\item{sinkends}{Boolean value related to whether data should be "sunk" to min and max values even if there are no measured values on which to base the interpolation.}
}
\value{
Interpolated data for the variable of interest.
}
\description{
Interpolates temperature data to populate missing depths at the date values for which data exist.
}
\examples{
interpolated_temps <- interpolateDepths(temp_profiles, date_col = "Date_LT", depth_col = "Depths", var = "Temp",
 interval = 0.1, min = 0, max = 15, sinkends = TRUE)
}
